+++ package.xml
@@ -1,7 +1,7 @@
 <package format="2">
-  <name>ros_destin</name>
+  <name>tensorflow_node</name>
   <version>0.0.1</version>
-  <description>Tensorflow implementation of DeSTIN</description>
+  <description>Persistent Tensorflow node</description>
   <maintainer email="mail@elggem.pub">Ralf Mayet</maintainer>
   <license>Unlicense</license>
