+++ scripts/daemon
@@ -7,37 +7,35 @@ import rospy
 import tensorflow as tf
 import numpy as np
 
-from destin import *
+from tensorflow_node import *
 
 from std_msgs.msg import Header
-from ros_destin.msg import DestinNodeState
+from tensorflow_node.msg import TFNodeState
 
 def str_to_class(str):
     return getattr(sys.modules[__name__], str)
 
 with tf.Session() as sess:
     # initialize ROS node
-    rospy.init_node('destin', anonymous=False, log_level=rospy.INFO)
-    rospy.loginfo("Destin ROS node launching")
+    rospy.init_node('tensorflow_daemon', anonymous=False, log_level=rospy.INFO)
+    rospy.loginfo("Tensorflow daemon ROS node launching")
 
-    # TODO Assertions: 
-    #   - inputlayer size and receptive field / stride fit together...
-    #   - ...?
+    config_prefix = "tensorflow_node"
 
     # initialize input layer from yaml
-    inputlayer_type = rospy.get_param("destin/inputlayer/type")
+    inputlayer_type = rospy.get_param(config_prefix+"/inputlayer/type")
     inputlayer_class = str_to_class(inputlayer_type)
-    inputlayer_params = rospy.get_param("destin/inputlayer/params")
+    inputlayer_params = rospy.get_param(config_prefix+"/inputlayer/params")
     inputlayer = inputlayer_class(**inputlayer_params)
 
     # initialize network from yaml
-    architecture_type = rospy.get_param("destin/architecture/type")
+    architecture_type = rospy.get_param(config_prefix+"/architecture/type")
     architecture_class = str_to_class(architecture_type)
-    architecture_params = rospy.get_param("destin/architecture/params")
+    architecture_params = rospy.get_param(config_prefix+"/architecture/params")
     architecture = architecture_class(sess, inputlayer, **architecture_params)
 
     # initialize summary writer
-    if (rospy.get_param("destin/publishing/summaries")):
+    if (rospy.get_param(config_prefix+"/publishing/summaries")):
         rospy.loginfo("recording summaries to " + SummaryWriter().get_summary_folder())
         # initialize summary writer with graph
         SummaryWriter().writer.add_graph(sess.graph)
@@ -45,10 +43,10 @@ with tf.Session() as sess:
 
     # initialize publishers for network
     publishers = {}
-    topic_name = rospy.get_param("destin/publishing/topic")
-    queue_size = rospy.get_param("/destin/inputlayer/params/batch_size")
+    topic_name = rospy.get_param(config_prefix+"/publishing/topic")
+    queue_size = rospy.get_param(config_prefix+"/inputlayer/params/batch_size")
     for node in architecture.nodes:
-        publishers[node.name] = rospy.Publisher('/'+topic_name+'/'+node.name, DestinNodeState, queue_size=queue_size)
+        publishers[node.name] = rospy.Publisher('/'+topic_name+'/'+node.name, TFNodeState, queue_size=queue_size)
 
     # main callback to evaluate architecture and publish states
     iteration = 0
@@ -67,7 +65,7 @@ with tf.Session() as sess:
 
             for state in ae_state:
                 # formulate message
-                msg = DestinNodeState()
+                msg = TFNodeState()
     
                 msg.header = Header()
                 msg.header.stamp = rospy.Time.now()
@@ -80,7 +78,7 @@ with tf.Session() as sess:
                 publishers[node.name].publish(msg)
         
         # publish summary output
-        if (rospy.get_param("destin/publishing/summaries")):
+        if (rospy.get_param(config_prefix+"/publishing/summaries")):
             summary_str = merged_summary_op.eval(feed_dict=feed_dict, session=sess)
             SummaryWriter().writer.add_summary(summary_str, iteration)
             SummaryWriter().writer.flush()
