+++ scripts/tensorboard
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-tensorboard --logdir=runs:`rosparam get destin/publishing/summary_folder` --port 6006 --reload_interval 5
+tensorboard --logdir=runs:`rosparam get tensorflow_node/publishing/summary_folder` --port 6006 --reload_interval 5
