+++ src/tensorflow_node/architectures/__init__.py
@@ -1,3 +1,3 @@
 from .architecture import NetworkArchitecture
-from .destin_arch import DestinArchitecture
+from .destin import DestinArchitecture
 from .handcoded_destin import HandcodedDestinArchitecture
