+++ src/tensorflow_node/nodes/node.py
