+++ src/tensorflow_node/nodes/stacked_autoencoder.py
