+++ src/tensorflow_node/tests/test_autoencoder.py
@@ -6,9 +6,9 @@ import logging as log
 import tensorflow as tf
 import numpy as np
 
-from destin import AutoEncoderNode
-from destin import SummaryWriter
-from destin import OpenCVInputLayer
+from tensorflow_node import AutoEncoderNode
+from tensorflow_node import SummaryWriter
+from tensorflow_node import OpenCVInputLayer
 
 
 # this tests very basic functionality of the autoencoder
